requirements.txt
google-generativeai>=0.4.0
opencv-python>=4.10.0
rich>=13.7
python-dotenv>=1.0
tenacity>=8.2
ffmpeg-python>=0.2       
